Simplificación
`main.cpp` en src está simplificado, puede hacer `CantidadLecturas`, pero no puede admitir operaciones concurrentes. Todo el código anterior está bajo `unused\main_prev.cpp`.
