--- src/Practica2Pt2.code-workspace
@@ -1,9 +1,0 @@
-{
-	"folders": [
-		{
-			"name": "Practica2Pt2",
-			"path": ".."
-		}
-	],
-	"settings": {}
-}
